feat!(plugin): Non async initial context (#169)
Removes `getInitialContext` from the plugin to a function as part of the library. This eliminates the need for `getInitialContext` from being asynchronous as it is being set on the window.

BREAKING CHANGE: This will break users currently using `Portals.getInitialContext`. It no longer returns a `Promise` and will return `undefined` if portalsInitialContext is not set on `window`.