feat: Update InitialContext (#184)
feat: Make InitialContext.value T | undefined. This is to support that
InitialContext will always be available when run in a portal context
with at least the name value defined. This now makes `getInitialContext` an
indicator to the web developers whether or not their code is being
executed inside of a Portal or not without the native team having to
provide any initialContext if it is not needed.

BREAKING CHANGE
This is an API breaking change since InitialContext<T> was initially
assumed to always have a value provided if it was present. This is now
changed so that we always provide access to the name of the portal to
the web code to avoid there being strong coupling to the name.